third_party/blink/web_tests/external/wpt/css/css-animations/CSSAnimation-effect.tentative-expected.txt

This is a testharness.js-based test.
PASS Setting a null effect on a running animation fires an animationend event
PASS Replacing an animation's effect with an effect that targets a different property should update both properties
PASS Replacing an animation's effect with a shorter one that should have already finished, the animation finishes immediately
PASS A play-pending animation's effect whose effect is replaced still exits the pending state
PASS CSS animation events are dispatched at the original element even after setting an effect with a different target element
PASS After replacing a finished animation's effect with a longer one it fires an animationstart event
FAIL Setting animation-composition sets the composite property on the effect assert_equals: expected "add" but got "replace"
PASS Replacing the effect of a CSSAnimation causes subsequent changes to corresponding animation-* properties to be ignored
Harness: the test ran to completion.
